Error sending the TrafficFlowObserved data model.
Hello, trying to send the example of the TrafficFlowObserved data model to the Context Broker, the Rest client was throwing me an error. in this commit the error is solved.

Error:
 ` {
         "error": "ParseError",
         "description": "Errors found in incoming JSON buffer"
} ` 
Cause: The comma in the attribute is missing:  `"capacity": 0.76`.
Solution: The comma is added to the end of the attribute.

Greetings.